레디스 추가

+++ server/src/server.ts
@@ -14,7 +14,6 @@ import {
   type PlanktonEatResponse,
   type ChatMessageSendResponse,
   type ChatMessageReceiveRequest,
-  type PlayerAttackResponse,
   type Species,
   type EvolveRequest,
   type NicknameRequest
@@ -26,6 +25,8 @@ import { SPECIES_ASSET } from "./constants/asset";
 import { getErrorMessage, getSuccessMessage } from "./message/message-handler";
 import { typeEnsure, recordEnsure } from "@/util/assert";
 import g from "@/types/global";
+import { error } from "console";
+import { getPlayer, setPlayer } from "./repository/redis";
 
 const dirname = path.resolve();
 const port: number = 3200; // 소켓 서버 포트
@@ -61,8 +62,16 @@ app.get("/", (req: Request, res: Response) => {
 });
 
 // 플레이어 위치 싱크
+// async function updatePlayerPositions(): Promise<void> {
+//   const playerList = await playerService.getPlayerList();
+//   sendToAll("others-position-sync", playerList);
+// }
+
+// 최초 실행 후 2초마다 updatePlayerPositions 함수를 호출
 // setInterval(() => {
-//   sendToAll("others-position-sync", playerService.getPlayerList());
+//   updatePlayerPositions().catch((error) => {
+//     console.error(error);
+//   });
 // }, 300);
 
 Container.set("width", 2688);
@@ -72,59 +81,52 @@ Container.set("planktonCnt", Math.floor(PLANKTON_SPAWN_LIST.length / 8));
 io.on("connection", (socket: Socket) => {
   const planktonManager = Container.get<PlanktonService>(PlanktonService);
 
-  if (g.playerList?.size === 0) {
+  if (playerService.count === 0) {
     planktonManager.initPlankton();
   }
 
   // 닉네임 확인
-  socket.on("nickname-validate", (nickname: NicknameRequest, callback) => {
+  socket.on("nickname-validate", async (nickname: NicknameRequest, callback) => {
     let validateResponse: ValidateRespone = {
       isSuccess: true,
       msg: getSuccessMessage("NICKNAME_VALIDATE_SUCCESS")
     };
-    try {
-      validateResponse = playerService.validateNickName(typeEnsure(nickname, "INVALID_INPUT"));
-    } catch (error: unknown) {
+    const result = await playerService.validateNickName(typeEnsure(nickname, "INVALID_INPUT"));
+
+    validateResponse = result;
+
+    if (!validateResponse.isSuccess) {
       validateResponse.isSuccess = false;
       validateResponse.msg = getErrorMessage(error);
-    } finally {
-      callback(validateResponse);
     }
+    callback(validateResponse);
   });
 
   // 참가자 본인 입장(소켓 연결)
-  socket.on("player-enter", (player: Player, callback) => {
+  socket.on("player-enter", async (player: Player, callback) => {
     const validResponse: ValidateRespone = {
       isSuccess: true,
       msg: getSuccessMessage("PLAYER_ENTER_SUCCESS")
     };
-    let gameStartReq: PlayerResponse = {
-      myInfo: player,
-      playerList: Array.from(g.playerList.values())
-    };
-
-    try {
-      if (playerService.validateNickName(typeEnsure(player, "INVALID_INPUT")).isSuccess) {
-        void socket.join(roomId);
-        gameStartReq = playerService.addPlayer(player, socket.id);
-      } else {
-        throw new Error("INVALID_INPUT");
-      }
-    } catch (error: unknown) {
-      validResponse.isSuccess = false;
-      validResponse.msg = getErrorMessage(error);
-    } finally {
-      callback(validResponse);
-
-      if (validResponse.isSuccess) {
+    const result = await playerService.validateNickName(typeEnsure(player, "INVALID_INPUT"));
+    if (result.isSuccess) {
+      void socket.join(roomId);
+      const addResult = await playerService.addPlayer(player, socket.id);
+      const gameStartReq: PlayerResponse | null = addResult;
+      if (gameStartReq !== null) {
         const planktonList: Plankton[] = Array.from(g.planktonList.values());
+        console.log(gameStartReq.myInfo);
         sendWithoutMe(socket, "player-enter", gameStartReq.myInfo);
         sendToMe(socket.id, "game-start", { ...gameStartReq, planktonList } satisfies GameStartData);
       }
+    } else {
+      validResponse.isSuccess = false;
+      validResponse.msg = getErrorMessage(error);
+      callback(validResponse);
     }
   });
   // 진화요청(Client→ Server)
-  socket.on("player-evolution", (data: EvolveRequest, callback) => {
+  socket.on("player-evolution", async (data: EvolveRequest, callback) => {
     let validateResponse: ValidateRespone = {
       isSuccess: true,
       msg: getSuccessMessage("PLAYER_EVOLVE_SUCCESS")
@@ -132,12 +134,12 @@ io.on("connection", (socket: Socket) => {
 
     try {
       recordEnsure(data, "INVALID_INPUT");
-      const beforeEvolvePlayer: Player = typeEnsure(g.playerList.get(data.playerId), "CANNOT_FIND_PLAYER");
+      const beforeEvolvePlayer: Player = typeEnsure(await getPlayer(data.playerId), "CANNOT_FIND_PLAYER");
       validateResponse = playerService.validateEvolution(data.speciesId, beforeEvolvePlayer);
       if (validateResponse.isSuccess) {
-        playerService.playerEvolution(data.speciesId, beforeEvolvePlayer);
+        await playerService.playerEvolution(data.speciesId, beforeEvolvePlayer);
         const { socketId, ...playerResponse } = beforeEvolvePlayer;
-        g.playerList.set(data.playerId, beforeEvolvePlayer);
+        await setPlayer(data.playerId, beforeEvolvePlayer);
         sendToMe(beforeEvolvePlayer.socketId, "player-status-sync", playerResponse);
       }
     } catch (error: unknown) {
@@ -151,44 +153,50 @@ io.on("connection", (socket: Socket) => {
   // game start
 
   // 플레이어 본인 위치 전송
-  socket.on("my-position-sync", (data: Player) => {
+  socket.on("my-position-sync", async (data: Player) => {
     try {
-      playerService.updatePlayerInfo(data);
-      const result = playerService.getPlayerList();
+      const result = await playerService.updatePlayerInfo(data);
+
       // 다른 플레이어에게 변경사항 알려줌
       sendWithoutMe(socket, "others-position-sync", result);
     } catch (error) {
-      // TO-DO: 플레이어가 서버에서 관리하지 않은 미검증된 사용자의 요청인 경우
-      // 처리 방법 상의가 필요합니다.
+      // 플레이어가 존재하지 않는 경우 퇴장 요청 날림
+      sendToAll("player-quit", data.playerId);
+      console.log(error);
     }
   });
 
   // 플레이어 본인 퇴장
-  socket.on("player-quit", (playerId: number) => {
-    playerService.deletePlayerByPlayerId(typeEnsure(playerId));
-    sendWithoutMe(socket, "player-quit", playerId);
-
+  socket.on("player-quit", async (playerId: number) => {
+    try {
+      await playerService.deletePlayerByPlayerId(typeEnsure(playerId));
+      sendWithoutMe(socket, "player-quit", playerId);
+    } catch (error) {
+      // 플레이어 삭제에 실패하면 에러 메세지
+      console.error(error);
+    }
     // playerId가 올바른 input이 아니라면 어떻게 해야할지
     // 논의가 필요합니다.
   });
 
   // 새로고침이나 창닫음으로 연결이 끊기는 경우
-  socket.on("disconnect", () => {
-    const result = playerService.deletePlayerBySocketId(socket.id);
+  socket.on("disconnect", async () => {
+    const result = await playerService.deletePlayerBySocketId(socket.id);
+
     if (result !== -1) {
       sendWithoutMe(socket, "player-quit", result);
     }
   });
 
   // 플랑크톤 섭취 이벤트
-  socket.on("plankton-eat", (data: { playerId: number; planktonId: number }, callback) => {
+  socket.on("plankton-eat", async (data: { playerId: number; planktonId: number }, callback) => {
     let result: PlanktonEatResponse = {
       isSuccess: true,
       msg: getSuccessMessage("EAT_PLANKTON_SUCCESS")
     };
     try {
       recordEnsure(data, "INVALID_INPUT");
-      result = planktonManager.eatedPlankton(data.planktonId, data.playerId);
+      result = await planktonManager.eatedPlankton(data.planktonId, data.playerId);
     } catch (error: unknown) {
       result.isSuccess = false;
       result.msg = getErrorMessage(error);
@@ -205,45 +213,58 @@ io.on("connection", (socket: Socket) => {
   });
 
   // 플레이어 간 공격
-  socket.on("player-crash", (data: PlayerCrashRequest, callback) => {
+  socket.on("player-crash", async (data: PlayerCrashRequest, callback) => {
     const validateResponse: ValidateRespone = {
       isSuccess: true,
       msg: getSuccessMessage("COLLISION_VALIDATE_SUCCESS")
     };
-    try {
-      playerService.isCrashValidate(data);
-    } catch (error: unknown) {
-      console.log(error);
-      validateResponse.isSuccess = false;
-      validateResponse.msg = getErrorMessage(error);
-    } finally {
-      callback(validateResponse);
+    // 요청이 왔는데 플레이어가 존하지않으면 나갔다고 처리
+    let isExist: boolean = true;
+    if (data.playerAId === undefined) {
+      isExist = false;
+      sendWithoutMe(socket, "player-quit", data.playerAId);
     }
+    if (data.playerBId === undefined) {
+      isExist = false;
+      sendWithoutMe(socket, "player-quit", data.playerBId);
+    }
+    if (isExist) {
+      try {
+        await playerService.isCrashValidate(data);
 
-    // 충돌 검증이 성공적인 경우만 공격 시도
-    if (validateResponse.isSuccess) {
-      const result: PlayerAttackResponse[] | undefined = playerService.attackPlayer(data);
+        if (validateResponse.isSuccess) {
+          const result = await playerService.attackPlayer(data);
+          if (result !== undefined && result.length === 2) {
+            // 플레이어 상태 정보 수정
+            for (const player of result) {
+              const mySocketId: string = player.socketId;
+              const { socketId, ...playerResponse } = player;
+              // 싱크 맞추는 부분에 대한 최적화 필요
+              sendToMe(mySocketId, "player-status-sync", playerResponse);
 
-      if (result !== undefined && result.length === 2) {
-        // 플레이어 상태 정보 수정
-        result.forEach((player) => {
-          const mySocketId: string = player.socketId;
-          const { socketId, ...playerResponse } = player;
-          sendToMe(mySocketId, "player-status-sync", playerResponse);
-
-          // 게임 오버인 경우
-          if (player.isGameOver) {
-            console.log("game-over");
-            sendToMe(player.socketId, "game-over", playerService.getGameOver(result));
-            sendWithoutMe(socket, "player-quit", player.playerId);
-            playerService.deletePlayerByPlayerId(player.playerId);
+              // 게임 오버인 경우
+              if (player.isGameOver) {
+                console.log("game-over");
+                const gameOverResponse = await playerService.getGameOver(result);
+                sendToMe(mySocketId, "game-over", gameOverResponse);
+                sendWithoutMe(socket, "player-quit", player.playerId);
+                await playerService.deletePlayerByPlayerId(player.playerId);
+              }
+            }
           }
-        });
+        }
+      } catch (error) {
+        console.log(error);
+
+        validateResponse.isSuccess = false;
+        validateResponse.msg = getErrorMessage(error);
+      } finally {
+        callback(validateResponse);
       }
     }
   });
 
-  socket.on("chat-message-send", (data: ChatMessageSendResponse) => {
+  socket.on("chat-message-send", async (data: ChatMessageSendResponse) => {
     const response: ValidateRespone = {
       isSuccess: false,
       msg: "유효하지 않은 플레이어입니다."
@@ -258,7 +279,7 @@ io.on("connection", (socket: Socket) => {
 
     try {
       recordEnsure(data, "INVALID_INPUT");
-      const sender: Player = typeEnsure(g.playerList.get(data.playerId), "CANNOT_FIND_PLAYER");
+      const sender: Player = typeEnsure(await getPlayer(data.playerId), "CANNOT_FIND_PLAYER");
       const targetSpecies: Species = typeEnsure(SPECIES_ASSET.get(sender.speciesId), "CANNOT_FIND_TIER");
 
       response.isSuccess = true;
@@ -276,7 +297,6 @@ io.on("connection", (socket: Socket) => {
       response.msg = getErrorMessage(error);
     } finally {
       // callback(response);
-
       if (response.isSuccess) {
         sendToAll("chat-message-receive", sendFormat);
       }
